Refactors NodeType into helpers
