+++ src/network/message.rs
@@ -14,7 +14,10 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{helpers::State, Environment, NodeType};
+use crate::{
+    helpers::{NodeType, State},
+    Environment,
+};
 use snarkos_storage::BlockLocators;
 use snarkvm::prelude::*;
