+++ src/network/prover.rs
@@ -14,7 +14,17 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{helpers::State, Data, Environment, LedgerReader, LedgerRequest, LedgerRouter, Message, NodeType, PeersRequest, PeersRouter};
+use crate::{
+    helpers::{NodeType, State},
+    Data,
+    Environment,
+    LedgerReader,
+    LedgerRequest,
+    LedgerRouter,
+    Message,
+    PeersRequest,
+    PeersRouter,
+};
 use snarkos_storage::{storage::Storage, ProverState};
 use snarkvm::dpc::prelude::*;
